Grid cell lenghts improved to variable size

+++ MonoGame.PortableUI/Common/GridLength.cs
@@ -4,13 +4,13 @@ namespace MonoGame.PortableUI.Common
     {
         public static GridLength Auto = new GridLength(1, GridLengthUnit.Auto);
 
-        public GridLength(int value) : this()
+        public GridLength(float value) : this()
         {
             Value = value;
             Unit = GridLengthUnit.Absolute;
         }
 
-        public GridLength(int value, GridLengthUnit unit) : this()
+        public GridLength(float value, GridLengthUnit unit) : this()
         {
             Value = value;
             Unit = unit;
